Always serialize call stack for ExecutionStackTree view
Previously, when expand_spans was enabled, the stack field was set to empty
array []. However, the ExecutionStackTree component requires the stack field
to display the call stack tree visualization.

This change ensures the stack is always serialized, regardless of expand_spans
setting, so that both the tree view and child spans can coexist.